Remove previous commit
commit bef751e2a25b9cb42f7741468374c71f28e8759d
I don't remember the logic for this commit but it was wrong

I noticed the issue when working with optim experiments on BRT
continuous Nishikawa data, with a biomass constraint

Oddly the total_stock was not the same at the end of the
optimization process and for a simulation run based on the
solution. Then with trials I noticed the total_stock varied
along the optimization process when it shouldn't have varied
(because optim only on q_sp_larvae)

The issue was associated to the definition 'fishing = true' in
the first block "READING CATCH DATA: C_obs, effort". With my
modification in the commit bef751..., this block was not
executed, so for the first time step, fishing was 0 (it is
redefined to 1 later in the code, so for the second time step
it was 1 again).

This prevented to compute Selectivity at the first time step
for the first optimization run, so Selectivity=0. But for the
second  optimization run, pop.Selectivity was no longer null
(=value from the last time step of the first optimization run)

What was confusing and made it really difficult to find the
issue and understand its consequences:
	- fishing is redefined when Precalrec_Calrec_adult for tag
	cohorts. So the issue with fishing=false was only visible
	for the first time step.
	- pop.Selectivity, if not redefined, is the one calculated
	at the end of the previous optimization run.