This is initial commit to new branch dev-skjmodel
This branch is the evolution of master branch and dev_larvae_dynpop branch, the latter being the significant enhancement of master branch by integrating the larval and other early life data into parameter estimation with the MLE approach. Compared to both origins, dev-skjmodel includes a series of major changes, either significantly impacting/improving model solutions and parameter estimation and/or involving substantial changes of adjoint functions, as well as multiple minor changes related to the code improvements. They all are listed and briefly explained in README.md. As the branch name states, it is the result of updating the reference skipjack model.