Constrain the use of tags by time at liberty
The use of short-lived (< time step) tags can bias the movement
parameters estimation because they are always moved entire time step.
Introduce two parameters: tags_tlib_min and tags_tlib_max to enable
flexible set-up in the parfile. Set the lower limit to 10 days and
allow user to increase it.
Besides, fix a small bug when the tags with 1 day at liberty are
released on the 31st of a month and attributed to the next time step
because of 30-days calendar. As a result these tags were ignored since
they were released post the time step of recapture.