Update dym library
This update includes a fix of function 'read.var.dym' to enable
extraction of a transect without loosing the 3D dimension of the output.
Also, some more checks and functionality added to the date extraction.
Namely, if either of t0.user or tfin.user are not specified, the data
from the first date or until the last date respectively will be
extracted. Also, function mult.dyms was added to the tookbox.