Fix indexing in writing spname_ageN.dym
Current writing of spname_ageN.dym is not correct: we have
spname_age2 = spname_larvae. Keep the age classes numbered
from 1 to A+ number, where '1' refers to  larvae at the end
of the time step and aged one month etc. For this, function
SaveCohortsDym is moved before Survival. Also, rename this
function to SaveDensityAtAgeDym and SaveCohorts to SaveRestart.