Parameterize HS's f_sst in early mortality
Parameterize HS's f_sst in early mortality

Add possibility to use HS's fsst in early mortality of ELM: it can either
use parameters a_sst_larvae and b_sst_larvae and so estimate them, or it
can be parameterized with fixed parameters, the XML line for this is the
following
	<hs_sst_func_coefs a_sst="6.35" b_sst="35.0" flag="1"/>
in section of <early_larvae_model>
If flag is set to 0 here, then a_sst_larvae and b_sst_larvae will be used.
Note, only fsst_type=1 (Gaussian) is currently implemented in early
mortality, other two types to be added if proven necessary here.
Note, XML files needs to be UPDATED following this change.